Add files via upload
This lab will take you through the steps of using the automated testing capabilities of the IBM Z Virtual Test Platform (VTP) to create a test at the transaction levels of a COBOL/CICS/ DB2 application. This is done by stubbing out CICS, IMS, DB2, DL/1 and MQ calls, enabling the program to be tested without a DB2 and CICS environment being active. In this lab you will record the transaction using an existing COBOL/CICS/DB2 application and write the recorded data into a flat file (playback file). Then you will make a change to the COBOL/CICS/DB2 program adding a bug and rerun the test that should catch the bug