Add files via upload
This lab will take you through the steps of using the automated unit testing (zUnit) capabilities of IBM Developer for z (IDz) to create a unit test case for a COBOL CICS/DB2 program. This enables a developer to test early without impacting other developers that share the same CICS environment and using batch jobs. In this lab you will record interaction with a COBOL CICS/DB2 program and run a test case where we add a bug in the program to be tested.